added first version of transferring resources between nodes

+++ Assets/Scripts/ResourceAmount.cs
@@ -2,9 +2,9 @@
 public class ResourceAmount
 {
     public Resource Resource;
-    public float Amount;
+    public int Amount;
 
-    public ResourceAmount(Resource resource, float amount)
+    public ResourceAmount(Resource resource, int amount)
     {
         Resource = resource;
         Amount = amount;
